nonlimber is working now

+++ cfftlog/cfftlog.h
@@ -4,6 +4,6 @@ typedef struct config {
 	int derivative;
 } config;
 
-void cfftlog(double *x, double *fx, long N, config *config, double ell, double *y, double *Fy);
+void cfftlog(double *x, double *fx, long N, config *config, int ell, double *y, double *Fy);
 
-void cfftlog_ells(double *x, double *fx, long N, config *config, double* ell, long Nell, double **y, double **Fy);
+void cfftlog_ells(double *x, double *fx, long N, config *config, int* ell, long Nell, double **y, double **Fy);
